Checked/approved Celia's Qadd; clarified comments
Qadd and Vadd treat the moulin volume change from creep/elastic deformation, which would change the water level.  We represent that as a change in water flux (Qin_compensated = Qin + Qadd) that we pass to subglacialsc, which calculates the new head.  Celia added these changes last week and I have now checked them, all seem correct to me, and added comments to clarify in a few places.  This is ready to merge to master.